Add a script to generate catalog mapping objects
