+++ modules/product/package.json
@@ -1,0 +1,13 @@
+{
+  "name": "product",
+  "version": "1.0.0",
+  "scripts": {
+    "generateMapping": "ts-node ./src/generateMappings.ts",
+    "test": "jest --group=-integration",
+    "it-test": "jest --group=integration"
+  },
+  "devDependencies": {
+    "ts-node": "^10.9.2",
+    "tsconfig-paths": "^4.2.0"
+  }
+}
